+++ springql-core/tests/feat_join.rs
@@ -56,7 +56,7 @@ fn run_and_drain(
     test_source_city_temperature: ForeignSource,
     test_sink: &ForeignSink,
 ) -> Vec<serde_json::Value> {
-    let _pipeline = apply_ddls(ddls, spring_config_default());
+    let _pipeline = apply_ddls(ddls, SpringConfig::default());
 
     test_source_trade.start(test_source_trade_input);
     test_source_city_temperature.start(test_source_city_temperature_input);